+++ __tests__/lib/lookerUtils.test.ts
@@ -34,8 +34,8 @@ describe("getExperimentLookerDashboardDate", () => {
 });
 
 describe("getLookerSubmissionTimestampDateFilter", () => {
-  it("returns the default date filter when startDate and endDate are null", () => {
-    const result = getLookerSubmissionTimestampDateFilter(null, null);
+  it("returns the default date filter when startDate and endDate are undefined", () => {
+    const result = getLookerSubmissionTimestampDateFilter();
 
     expect(result).toEqual("30 day ago for 30 day");
   });
@@ -64,5 +64,19 @@ describe("getLookerSubmissionTimestampDateFilter", () => {
     const result = getLookerSubmissionTimestampDateFilter(startDate, endDate);
 
     expect(result).toEqual("2024-05-08 to 3024-06-10");
+  });
+
+  it("returns a date filter from the startDate to endDate when startDate and endDate are defined and the message is completed", () => {
+    const startDate = "2024-05-08";
+    const endDate = "2024-06-10";
+    const isCompleted = true;
+
+    const result = getLookerSubmissionTimestampDateFilter(
+      startDate,
+      endDate,
+      isCompleted,
+    );
+
+    expect(result).toEqual("2024-05-08 to 2024-06-10");
   });
 });