+++ lib/lookerUtils.ts
@@ -25,20 +25,41 @@ export function getExperimentLookerDashboardDate(
 }
 
 /**
- * @returns the submission timestamp date filter to be used in the Looker dashboard links
+ * @param startDate The date that the Looker submission timestamp date filter
+ *        will start at if it's defined.
+ * @param endDate Either the calculated proposed end date or the actual end
+ *        date that the Looker submission timestamp date filter will end at
+ *        if it's defined. endDate should be one day ahead of the actual date
+ *        because Looker date filters don't include the end date.
+ * @param isCompleted True if the experiment is completed.
+ * @returns The submission timestamp date filter to be used in the Looker dashboard links.
  */
 export function getLookerSubmissionTimestampDateFilter(
   startDate?: string | null,
   endDate?: string | null,
+  isCompleted?: boolean,
 ): string {
-  // Showing the last 30 complete days to ensure the dashboard isn't including today which has no data yet
-  let submission_timestamp_date = "30 day ago for 30 day";
-
-  if (startDate && endDate && new Date() < new Date(endDate)) {
-    submission_timestamp_date = `${startDate} to ${endDate}`;
+  if (isCompleted && startDate && endDate) {
+    // This case covers completed experiments with defined startDate and
+    // endDate from the Experimenter API.
+    return `${startDate} to ${endDate}`;
+  } else if (
+    !isCompleted &&
+    startDate &&
+    endDate &&
+    new Date() < new Date(endDate)
+  ) {
+    // This case covers experiments that haven't reached their proposed end date.
+    return `${startDate} to ${endDate}`;
   } else if (startDate) {
-    submission_timestamp_date = `${startDate} to today`;
-  }
+    // This case covers experiments that are still ongoing past their proposed
+    // end date.
+    return `${startDate} to today`;
+  } else {
+    // This case covers any messages with undefined startDate and endDate.
 
-  return submission_timestamp_date;
+    // Showing the last 30 complete days to ensure the dashboard isn't
+    // including today which has no data yet.
+    return "30 day ago for 30 day";
+  }
 }