+++ lib/nimbusRecipe.ts
@@ -10,6 +10,7 @@ import {
   getProposedEndDate,
   MESSAGING_EXPERIMENTS_DEFAULT_FEATURES,
   _substituteLocalizations,
+  formatDate,
 } from "../lib/experimentUtils.ts";
 import { getExperimentLookerDashboardDate } from "./lookerUtils.ts";
 
@@ -50,6 +51,7 @@ function _branchInfoHasMicrosurvey(branchInfo: BranchInfo): boolean {
 
 type NimbusRecipeType = {
   _rawRecipe: NimbusExperiment;
+  _isCompleted: boolean;
 
   getRecipeInfo(): RecipeInfo;
   getRecipeOrBranchInfos(): RecipeOrBranchInfo[];
@@ -63,9 +65,11 @@ type NimbusRecipeType = {
 
 export class NimbusRecipe implements NimbusRecipeType {
   _rawRecipe;
+  _isCompleted;
 
-  constructor(recipe: NimbusExperiment) {
+  constructor(recipe: NimbusExperiment, isCompleted: boolean = false) {
     this._rawRecipe = recipe;
+    this._isCompleted = isCompleted;
   }
 
   /**
@@ -240,6 +244,10 @@ export class NimbusRecipe implements NimbusRecipeType {
       branchInfo.nimbusExperiment.startDate,
       branchInfo.nimbusExperiment.proposedDuration,
     );
+    let formattedEndDate;
+    if (branchInfo.nimbusExperiment.endDate) {
+      formattedEndDate = formatDate(branchInfo.nimbusExperiment.endDate, 1);
+    }
     branchInfo.ctrDashboardLink = getDashboard(
       branch.template,
       branchInfo.id,
@@ -247,7 +255,8 @@ export class NimbusRecipe implements NimbusRecipeType {
       branchInfo.nimbusExperiment.slug,
       branch.slug,
       branchInfo.nimbusExperiment.startDate,
-      proposedEndDate,
+      branchInfo.nimbusExperiment.endDate ? formattedEndDate : proposedEndDate,
+      this._isCompleted,
     );
     // if (!feature.value.content) {
     //   console.log("v.content is null");